Monitorando a amazônia e almost shortest path done

+++ spoj/Almost_Shortest_Path/Almost_Shortest_Path.cpp
@@ -4,18 +4,18 @@
 typedef long long ll;
 using namespace std;
 
-void deleteConnections(vector<vector<pair<ll,ll> > >&adj, vector<ll>temp){
+void deleteConnections(vector<vector<pair<ll,ll> > >&adj, vector<ll>&temp){
     for(ll i = 0; i < temp.size() - 1; i++){
-        for(ll j = 0; j < adj[i].size(); j++){
-            if(adj[i][j].first == temp[i+1]){
-                adj[i].erase(adj[i].begin() + j);
+        for(ll j = 0; j < adj[temp[i]].size(); j++){
+            if(adj[temp[i]][j].first == temp[i+1]){
+                adj[temp[i]].erase(adj[temp[i]].begin() + j);
                 break;
             }
         }
     }
 }
 
-void djk(vector<vector<pair<ll,ll> > >&adj , ll s, ll d, vector<ll>&dist, ll mnsize){
+void djk(vector<vector<pair<ll,ll> > >&adj , ll s, ll d, vector<ll>&dist, ll mnsize, int todelete){
     set<pair<pair<ll,ll>, vector<ll> > > st;
     vector<ll>temp;
     dist[s] = 0;
@@ -23,16 +23,16 @@ void djk(vector<vector<pair<ll,ll> > >&adj , ll s, ll d, vector<ll>&dist, ll mns
     st.insert(make_pair(make_pair(0,s), temp));
     while(!st.empty()){
         ll no = (*st.begin()).first.second;
-        ll custo = (*st.begin()).first.second;
+        ll custo = (*st.begin()).first.first;
         vector<ll>temp = (*st.begin()).second;
 
         st.erase(st.begin());
-        if(no == d && custo == mnsize){
+        if(no == d && custo == mnsize && todelete){
             deleteConnections(adj, temp);
         }
 
         for(auto it : adj[no]){
-            if(custo + it.second < dist[it.first]){
+            if(custo + it.second <= dist[it.first]){
                 dist[it.first] = custo + it.second;
                 temp.push_back(it.first);
                 st.insert(make_pair(make_pair(dist[it.first], it.first), temp));
@@ -60,15 +60,15 @@ int main(){
             adj[a].push_back(make_pair(b,c));
         }
 
-        vector<ll>dist(n+1,1e18);
-        djk(adj,s,d,dist,0);
+        vector<ll>dist(n+1,1e18); 
+        djk(adj,s,d,dist,0, 0); //encontra o menor valor
 
         ll mnsize = dist[d];
-        dist.assign(n+1,1e18);
-        djk(adj,s,d,dist,mnsize);
+        dist.assign(n+1,1e18); 
+        djk(adj,s,d,dist,mnsize,1); //apaga os menores caminhos
 
         dist.assign(n+1,1e18);
-        djk(adj,s,d,dist,0);
+        djk(adj,s,d,dist,0, 0); // encontra o almost shortest
 
         if(dist[d] == 1e18){
             cout << -1 << endl;
